requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
pymongo==4.6.1
python-dotenv==1.0.1
transformers==4.36.2
torch==2.1.2
requests==2.31.0
beautifulsoup4==4.12.3
readability-lxml==0.8.1
trafilatura==1.6.4
